@@ -23,78 +23,72 @@
   "scripts": {
     "start": "webpack serve --env APP_ENV=dev",
     "build": "webpack build --env APP_ENV=prod",
-    "debug": "webpack build --env APP_ENV=debug",
-    "tsc": "tsc --noemit",
-    "postcss": "postcss src/**/*.{css,less,scss} --base src --dir build",
-    "babel": "babel src --extensions .js,.jsx,.ts,.tsx --out-dir lib",
-    "eslint": "eslint src/**/*.{js,jsx,ts,tsx}",
-    "stylelint": "stylelint src/**/*.{css,less,scss}",
-    "prettier": "prettier src/**/* --check",
-    "lint": "npm run eslint && npm run stylelint && npm run prettier",
+    "lint": "sh scripts/lint.sh",
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@babel/cli": "7.x",
-    "@babel/core": "7.x",
-    "@babel/plugin-transform-runtime": "7.x",
-    "@babel/preset-env": "7.x",
-    "@babel/preset-react": "7.x",
-    "@babel/preset-typescript": "7.x",
-    "@commitlint/cli": "16.x",
-    "@commitlint/config-conventional": "16.x",
-    "@types/lodash": "4.x",
-    "@types/node": "20.x",
-    "@types/react": "17.x",
-    "@types/react-dom": "17.x",
-    "@types/react-router-dom": "5.x",
-    "@typescript-eslint/eslint-plugin": "5.x",
-    "@typescript-eslint/parser": "5.x",
-    "babel-loader": "8.x",
-    "copy-webpack-plugin": "10.x",
-    "css-loader": "6.x",
-    "eslint": "8.x",
-    "eslint-config-prettier": "8.x",
-    "eslint-plugin-react": "7.x",
-    "html-webpack-plugin": "5.x",
-    "husky": "7.x",
-    "less": "4.x",
-    "less-loader": "10.x",
-    "lint-staged": "12.x",
-    "lodash": "4.x",
-    "mini-css-extract-plugin": "2.x",
-    "postcss": "8.x",
-    "postcss-cli": "9.x",
-    "postcss-less": "6.x",
-    "postcss-loader": "6.x",
-    "postcss-preset-env": "7.x",
-    "postcss-scss": "4.x",
-    "prettier": "2.x",
-    "sass": "1.x",
-    "sass-loader": "12.x",
-    "style-loader": "3.x",
-    "stylelint": "14.x",
-    "stylelint-config-prettier": "9.x",
-    "stylelint-config-standard": "24.x",
-    "thread-loader": "4.x",
-    "typescript": "4.x",
-    "webpack": "5.x",
-    "webpack-bundle-analyzer": "4.x",
-    "webpack-cli": "4.x",
-    "webpack-dev-server": "4.x",
-    "webpack-merge": "5.x"
+    "@babel/cli": "~7.23.9",
+    "@babel/core": "~7.24.0",
+    "@babel/plugin-transform-runtime": "~7.24.0",
+    "@babel/preset-env": "~7.24.0",
+    "@babel/preset-react": "~7.23.3",
+    "@babel/preset-typescript": "~7.23.3",
+    "@commitlint/cli": "~19.0.3",
+    "@commitlint/config-conventional": "~19.0.3",
+    "@types/lodash": "~4.14.202",
+    "@types/node": "~20.11.24",
+    "@types/react": "~18.2.61",
+    "@types/react-dom": "~18.2.19",
+    "@types/react-redux": "~7.1.33",
+    "@types/react-router-dom": "~5.3.3",
+    "@typescript-eslint/eslint-plugin": "~7.1.0",
+    "@typescript-eslint/parser": "~7.1.0",
+    "babel-loader": "~9.1.3",
+    "copy-webpack-plugin": "~12.0.2",
+    "css-loader": "~6.10.0",
+    "eslint": "~8.57.0",
+    "eslint-config-prettier": "~9.1.0",
+    "eslint-plugin-react": "~7.34.0",
+    "html-webpack-plugin": "~5.6.0",
+    "husky": "~9.0.11",
+    "less": "~4.2.0",
+    "less-loader": "~12.2.0",
+    "lint-staged": "~15.2.2",
+    "lodash": "~4.17.21",
+    "mini-css-extract-plugin": "~2.8.1",
+    "postcss": "~8.4.35",
+    "postcss-cli": "~11.0.0",
+    "postcss-less": "~6.0.0",
+    "postcss-loader": "~8.1.1",
+    "postcss-preset-env": "~9.4.0",
+    "postcss-scss": "~4.0.9",
+    "prettier": "~3.2.5",
+    "sass": "~1.71.1",
+    "sass-loader": "~14.1.1",
+    "style-loader": "~3.3.4",
+    "stylelint": "~16.2.1",
+    "stylelint-config-prettier": "~9.0.5",
+    "stylelint-config-standard": "~36.0.0",
+    "thread-loader": "~4.0.2",
+    "typescript": "~5.3.3",
+    "webpack": "~5.90.3",
+    "webpack-bundle-analyzer": "~4.10.1",
+    "webpack-cli": "~5.1.4",
+    "webpack-dev-server": "~5.0.2",
+    "webpack-merge": "~5.10.0"
   },
   "dependencies": {
-    "@ant-design/icons": "4.x",
-    "@babel/runtime": "7.x",
-    "antd": "4.x",
-    "axios": "0.x",
-    "core-js": "3.x",
-    "react": "18.x",
-    "react-dom": "18.x",
-    "react-redux": "7.x",
-    "react-router-dom": "5.x",
-    "redux": "4.x",
-    "redux-thunk": "2.x",
-    "regenerator-runtime": "0.x"
+    "@ant-design/icons": "~5.3.1",
+    "@babel/runtime": "~7.24.0",
+    "antd": "~5.15.0",
+    "axios": "~1.6.7",
+    "core-js": "~3.36.0",
+    "react": "~18.2.0",
+    "react-dom": "~18.2.0",
+    "react-redux": "~9.1.0",
+    "react-router-dom": "~6.22.2",
+    "redux": "~5.0.1",
+    "redux-thunk": "~3.1.0",
+    "regenerator-runtime": "~0.14.1"
   }
 }
