@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOMClient from 'react-dom/client';
-import { HashRouter } from 'react-router-dom';
 
 import './index.css';
 import App from './App';
@@ -12,9 +11,7 @@ import App from './App';
  */
 ReactDOMClient.createRoot(document.getElementById('root')).render(
   // <React.StrictMode>
-  <HashRouter>
-    <App />
-  </HashRouter>
+  <App />
   // </React.StrictMode>
 );
 
